Move shared search logic into `Homebrew::Search` module.
